dashboard-weather/weather_function/requirements.txt
google-cloud-firestore
google-cloud-storage
functions-framework
